siegewar - script bug fixes

+++ src/com/palmergames/bukkit/towny/TownySettings.java
@@ -2627,8 +2627,8 @@ public class TownySettings {
 		return getInt(ConfigNodes.WAR_SIEGE_TIMER_TICK_INTERVAL_SECONDS);
 	}
 
-	public static int getWarSiegeSiegeCooldownNewTownsHours() {
-		return getInt(ConfigNodes.WAR_SIEGE_SIEGE_COOLDOWN_NEW_TOWNS_HOURS);
+	public static double getWarSiegeSiegeCooldownNewTownsHours() {
+		return getDouble(ConfigNodes.WAR_SIEGE_SIEGE_COOLDOWN_NEW_TOWNS_HOURS);
 	}
 	public static double getWarSiegeSiegeCooldownModifier() {
 		return getDouble(ConfigNodes.WAR_SIEGE_SIEGE_COOLDOWN_MODIFIER);
